+++ connectivity_matrix.h
@@ -6,7 +6,7 @@
 class ConnectivityMatrix {
     public:
         ConnectivityMatrix (Layer from_layer, Layer to_layer,
-            bool plastic, double max_weight);
+            bool plastic, float max_weight);
 
         void build();
 
@@ -18,16 +18,16 @@ class ConnectivityMatrix {
         int to_size;
         int sign;
         bool plastic;
-        double max_weight;
-        double* mData;
+        float max_weight;
+        float* mData;
 
-        void randomize(bool self_connected, double max_weight);
+        void randomize(bool self_connected, float max_weight);
 
-        double& operator()(int i, int j) {
+        float& operator()(int i, int j) {
             return this->mData[i * to_size + j];
         }
 
-        double operator()(int i, int j) const {
+        float operator()(int i, int j) const {
             return this->mData[i * to_size + j];
         }
